Move discover filter fragment to streaming package.
