+++ app/src/main/res/layout/fragment_shows_discover_filter.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".ui.search.ShowsDiscoverFilterFragment">
+    tools:context=".streaming.DiscoverFilterFragment">
 
     <Button
         android:id="@+id/buttonWatchRegion"
